Move ffi to different location, use different submodule
